Move althrd_setname to its own source
