+++ alc/backends/opensl.cpp
@@ -35,6 +35,7 @@
 
 #include "albit.h"
 #include "alnumeric.h"
+#include "althrd_setname.h"
 #include "core/device.h"
 #include "core/helpers.h"
 #include "core/logging.h"
