+++ alc/backends/sndio.cpp
@@ -32,11 +32,11 @@
 #include <vector>
 
 #include "alnumeric.h"
+#include "althrd_setname.h"
 #include "core/device.h"
 #include "core/helpers.h"
 #include "core/logging.h"
 #include "ringbuffer.h"
-#include "threads.h"
 
 #include <sndio.h>
