+++ alc/backends/winmm.cpp
@@ -39,6 +39,7 @@
 #include <functional>
 
 #include "alnumeric.h"
+#include "althrd_setname.h"
 #include "core/device.h"
 #include "core/helpers.h"
 #include "core/logging.h"
